# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: nepapi
  version: "1.0.0"
  description: |
    Public JSON API for Nepal's public data: daily NRB foreign exchange
    rates and Bikram Sambat calendar conversion.

    Everything under /v1/ is a contract: fields may be added, never
    renamed, removed, or retyped. Rate limit: 2 requests/second
    sustained per IP, bursts up to 30 (429 + Retry-After beyond that).
  license:
    name: MIT
    url: https://github.com/prashantkoirala465/nepapi/blob/main/LICENSE
  contact:
    url: https://github.com/prashantkoirala465/nepapi

paths:
  /v1/forex/latest:
    get:
      operationId: forexLatest
      summary: All rates for the most recent published date
      responses:
        "200":
          description: Latest published rates
          content:
            application/json:
              schema:
                type: object
                required: [date, rates]
                properties:
                  date:
                    type: string
                    format: date
                    example: "2026-07-11"
                  rates:
                    type: array
                    items:
                      $ref: "#/components/schemas/ForexRate"
        "404":
          description: No rates stored yet (empty database)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"

  /v1/forex/rates:
    get:
      operationId: forexRates
      summary: Historical rates for a date range
      parameters:
        - name: from
          in: query
          required: true
          schema: {type: string, format: date}
          example: "2026-07-01"
        - name: to
          in: query
          required: true
          description: Inclusive; at most 366 days after `from`.
          schema: {type: string, format: date}
          example: "2026-07-11"
        - name: currency
          in: query
          required: false
          description: ISO 4217 alpha-3 filter; omit for all currencies.
          schema: {type: string}
          example: USD
      responses:
        "200":
          description: Rates in the range, ordered by date then currency
          content:
            application/json:
              schema:
                type: object
                required: [from, to, count, rates]
                properties:
                  from: {type: string, format: date}
                  to: {type: string, format: date}
                  count: {type: integer}
                  rates:
                    type: array
                    items:
                      $ref: "#/components/schemas/ForexRate"
        "400":
          description: Bad or missing parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"

  /v1/calendar/convert:
    get:
      operationId: calendarConvert
      summary: Convert between Gregorian (AD) and Bikram Sambat (BS) dates
      description: |
        Supported range: 1970–2090 BS (1913-04-13 to 2034-04-13 AD).
        Conversion is table-driven; out-of-range or impossible dates
        return 400 with the supported range in the message.
      parameters:
        - name: date
          in: query
          required: true
          schema: {type: string, pattern: '^\d{4}-\d{2}-\d{2}$'}
          example: "2026-07-11"
        - name: from
          in: query
          required: false
          description: Calendar of `date`. Defaults to `ad`.
          schema: {type: string, enum: [ad, bs], default: ad}
      responses:
        "200":
          description: The date in both calendars
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertResponse"
        "400":
          description: Bad date, bad `from`, or out of supported range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/holidays:
    get:
      operationId: holidays
      summary: Nepal's national public holidays for one BS year
      description: |
        Hand-curated dataset, one file per BS year (currently 2082 and
        2083). Islamic holidays are excluded until their moon-sighted
        dates are confirmed; see the dataset's attribution notes in the
        repository. Defaults to the current BS year.
      parameters:
        - name: year
          in: query
          required: false
          description: BS year. Defaults to the current BS year.
          schema: {type: integer}
          example: 2082
      responses:
        "200":
          description: Holidays in date order
          content:
            application/json:
              schema:
                type: object
                required: [year, count, holidays]
                properties:
                  year: {type: integer, example: 2082}
                  count: {type: integer, example: 34}
                  holidays:
                    type: array
                    items:
                      $ref: "#/components/schemas/Holiday"
        "400":
          description: Malformed year
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Year outside the curated dataset; message lists available years
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"

  /v1/health:
    get:
      operationId: health
      summary: Liveness — process is up, checks nothing else
      responses:
        "200":
          description: Alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string, const: ok}

  /v1/ready:
    get:
      operationId: ready
      summary: Readiness — database answers within 2s
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string, const: ready}
        "503":
          description: Database unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/openapi.yaml:
    get:
      operationId: openapiSpec
      summary: This specification
      responses:
        "200":
          description: OpenAPI 3.1 spec
          content:
            application/yaml: {}

  /docs:
    get:
      operationId: docs
      summary: Human-readable API documentation
      responses:
        "200":
          description: Documentation page
          content:
            text/html: {}

components:
  schemas:
    ForexRate:
      type: object
      required: [date, iso3, name, unit, buy, sell]
      properties:
        date:
          type: string
          format: date-time
          example: "2026-07-11T00:00:00Z"
        iso3:
          type: string
          description: ISO 4217 alpha-3 currency code
          example: USD
        name:
          type: string
          example: U.S. Dollar
        unit:
          type: integer
          description: >
            Rates are NPR per this many units of the currency. Check it —
            NRB quotes INR per 100 and JPY per 10.
          example: 1
        buy:
          type: string
          description: Decimal string with NRB's exact digits; parse with a decimal type, not a float.
          example: "152.23"
        sell:
          type: string
          description: Decimal string with NRB's exact digits.
          example: "152.83"

    ConvertResponse:
      type: object
      required: [bs, ad, weekday]
      properties:
        bs:
          type: object
          required: [year, month, day, month_name, month_name_ne]
          properties:
            year: {type: integer, example: 2083}
            month: {type: integer, example: 3}
            day: {type: integer, example: 27}
            month_name: {type: string, example: Ashadh}
            month_name_ne: {type: string, example: असार}
        ad:
          type: string
          format: date
          example: "2026-07-11"
        weekday:
          type: string
          example: Saturday

    Holiday:
      type: object
      required: [bs, ad, weekday, name, name_ne]
      properties:
        bs:
          type: object
          required: [year, month, day]
          properties:
            year: {type: integer, example: 2082}
            month: {type: integer, example: 6}
            day: {type: integer, example: 16}
        ad:
          type: string
          format: date
          example: "2025-10-02"
        weekday:
          type: string
          example: Thursday
        name:
          type: string
          example: Vijaya Dashami
        name_ne:
          type: string
          example: विजयादशमी
        note:
          type: string
          description: Present when the holiday applies to specific districts (e.g. the two-day Holi split).
          example: hilly districts

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "invalid or missing 'from' (want YYYY-MM-DD)"

  responses:
    RateLimited:
      description: Rate limit exceeded — honor the Retry-After header
      headers:
        Retry-After:
          schema: {type: string}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Internal:
      description: Our bug — details are in server logs, not the response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"